+++ src/main/java/qna/deletehistory/DeleteHistoryRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.deletehistory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.ContentType;
 
 import java.util.List;
