+++ src/main/java/qna/deletehistory/DeleteHistoryService.java
@@ -1,10 +1,8 @@
-package qna.service;
+package qna.deletehistory;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import qna.domain.DeleteHistory;
-import qna.domain.DeleteHistoryRepository;
 
 import java.util.List;
