+++ src/main/java/qna/question/QuestionRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.user.User;
 
 import java.util.List;
 import java.util.Optional;
