+++ src/test/java/qna/domain/UserRepositoryTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.UnAuthorizedException;
+import qna.user.User;
+import qna.user.UserRepository;
 
 import java.util.Optional;
